ai_projects/citation-guardrail-engine/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.0.0
openai>=1.30.0
python-dotenv>=1.0.0
pytest>=9.0.0
pytest-asyncio>=1.0.0
